app/requirements.txt
fastapi
uvicorn
python-dotenv
openai
streamlit
requests
datetime
